# swagger.yaml
---
swagger: "2.0"
info:
  description: "Annotation API (Using W3C Open Annotation Model (not protocol)). Supporting\
    \ the Annotation of content with concepts."
  version: "v0.0.1"
  title: "Annotation API"
  termsOfService: "http://www.yourterms.com"
  contact:
    name: "Jem Rayfield"
    url: "http://ontotext.com"
    email: "jem.rayfield@ontotext.com"
host: "localhost:8089"
basePath: "/annotations"
tags:
- name: "Annotation API"
schemes:
- "http"
- "https"
paths:
  /annotations:
    get:
      tags:
      - "Annotation API"
      summary: "Finds Annotations, by contentId if required, pageable with offsets"
      description: ""
      operationId: "annotations"
      produces:
      - "application/json"
      parameters:
      - name: "contentId"
        in: "query"
        description: "Target content Id"
        required: false
        type: "string"
        default: "963db081-4200-430f-89dc-5756ca8edf04"
      - name: "max"
        in: "query"
        description: "The number of records to display per page"
        required: false
        type: "integer"
        default: 10
        format: "int32"
      - name: "offset"
        in: "query"
        description: "The page to be returned"
        required: false
        type: "integer"
        default: 0
        format: "int32"
      responses:
        200:
          description: "Success"
          headers:
            X-Cache:
              type: "boolean"
              description: "Explains whether or not a cache was used"
            ETag:
              type: "string"
              description: "Annotation ETag for cache control"
            Vary:
              type: "string"
              description: "Make sure proxies cache by Vary"
            Link:
              type: "string"
              description: "Linked Data Platform resource type"
            Allow:
              type: "string"
              description: "CORS Allowed Methods"
        400:
          description: "Invalid Content Id supplied"
  /annotations/asynch:
    post:
      tags:
      - "Annotation API"
      summary: "Asynchronous publication/creation of annotations"
      description: ""
      operationId: "createAnnotationAsynch"
      consumes:
      - "application/ld+json; profile=\"http://www.w3.org/ns/anno.jsonld\""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Annotation to create"
        required: true
        schema:
          type: "string"
      responses:
        202:
          description: "Processing"
          headers:
            X-Cache:
              type: "boolean"
              description: "Explains whether or not a cache was used"
            ETag:
              type: "string"
              description: "Annotation ETag for cache control"
            Vary:
              type: "string"
              description: "Make sure proxies cache by Vary"
            Allow:
              type: "string"
              description: "CORS Allowed Methods"
        400:
          description: "Invalid Annotations supplied"
  /annotations/asynch/{processId}:
    get:
      tags:
      - "Annotation API"
      summary: "Asynchronous publication/creation annotation status"
      description: ""
      operationId: "statusAnnotationAsynch"
      produces:
      - "application/json"
      parameters:
      - name: "processId"
        in: "query"
        description: "Asynchronous Transaction IUd"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          headers:
            X-Cache:
              type: "boolean"
              description: "Explains whether or not a cache was used"
            ETag:
              type: "string"
              description: "Annotation ETag for cache control"
            Vary:
              type: "string"
              description: "Make sure proxies cache by Vary"
            Allow:
              type: "string"
              description: "CORS Allowed Methods"
        400:
          description: "Invalid ProcessId supplied"
  /annotations/{annotationId}:
    get:
      tags:
      - "Annotation API"
      summary: "Finds Annotation by id"
      description: ""
      operationId: "annotation"
      produces:
      - "application/ld+json; profile=\"http://www.w3.org/ns/anno.jsonld\""
      parameters:
      - name: "annotationId"
        in: "path"
        description: "Annotation Id to be retrieved. Must be a valid UUID"
        required: true
        type: "string"
        default: "21b4a142-af56-4dce-9f6c-0a6875933353"
      - name: "X-Request-ID"
        in: "header"
        description: "Transaction Id"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          headers:
            X-Cache:
              type: "boolean"
              description: "Explains whether or not a cache was used"
            ETag:
              type: "string"
              description: "Annotation ETag for cache control"
            Vary:
              type: "string"
              description: "Make sure proxies cache by Vary"
            Link:
              type: "string"
              description: "Linked Data Platform resource type"
            Allow:
              type: "string"
              description: "CORS Allowed Methods"
        400:
          description: "Invalid Annotation Id supplied"
        404:
          description: "Annotation not found"